backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
python-jose[cryptography]==3.3.0
passlib==1.7.4
docker>=7.1.0
requests<2.32.0
python-multipart==0.0.6
httpx==0.26.0
apscheduler==3.10.4
psutil==5.9.8
